HTTP/XML Print Server for Adafruit Thermal Printer
This is a Python/Flask-based HTTP server that uses the Adafruit Thermal
Printer library to communicate with a thermal receipt printer. The
server accepts formatted XML by HTTP POST, and parses it to extract
commands for the printer. This makes it possible to "push" content to
the printer (e.g., running on a Raspberry Pi), which opens up a number
of cloud-based automation options.